+++ provide.go
@@ -1,6 +1,9 @@
 package pluginfx
 
 import (
+	"os"
+	"path/filepath"
+
 	"go.uber.org/fx"
 )
 
@@ -20,7 +23,8 @@ type P struct {
 	// placed into the fx.App regardless of the values of Name and Group.
 	Anonymous bool
 
-	// Path is the plugin's path.  This field is required.
+	// Path is the plugin's path.  This field is required.  Variables are expanded
+	// via os.ExpandEnv.
 	Path string
 
 	// Symbols describes the optional set of functions exported by the plugin to be
@@ -54,7 +58,7 @@ type P struct {
 //   )
 func (p P) Provide() fx.Option {
 	var options []fx.Option
-	plugin, err := Open(p.Path)
+	plugin, err := Open(os.ExpandEnv(p.Path))
 
 	if err == nil {
 		options = append(options, p.Symbols.Load(plugin))
@@ -95,11 +99,16 @@ type S struct {
 	// field is unset, the loaded plugins are not added as components.
 	Group string
 
-	// Paths are the plugin paths to load.
+	// Paths are the plugin paths to load.  Each of these paths may be a filesystem glob,
+	// in which case all matching files are loaded as plugins.  Variable expansion is also
+	// done on each element via os.ExpandEnv.
 	Paths []string
 
+	// Symbols are the symbols to be loaded from each loaded plugin.
 	Symbols Symbols
 
+	// Lifecycle describes the symbols from each loaded plugin to be bound to the
+	// enclosing application.
 	Lifecycle Lifecycle
 }
 
@@ -109,16 +118,24 @@ type S struct {
 func (s S) Provide() fx.Option {
 	var options []fx.Option
 	for _, path := range s.Paths {
-		options = append(options,
-			P{
-				Group:     s.Group,
-				Anonymous: len(s.Group) == 0,
-				Path:      path,
+		matches, err := filepath.Glob(os.ExpandEnv(path))
+		if err != nil {
+			options = append(options, fx.Error(err))
+			continue
+		}
 
-				Symbols:   s.Symbols,
-				Lifecycle: s.Lifecycle,
-			}.Provide(),
-		)
+		for _, match := range matches {
+			options = append(options,
+				P{
+					Group:     s.Group,
+					Anonymous: len(s.Group) == 0,
+					Path:      match,
+
+					Symbols:   s.Symbols,
+					Lifecycle: s.Lifecycle,
+				}.Provide(),
+			)
+		}
 	}
 
 	return fx.Options(options...)