+++ provide_test.go
@@ -2,6 +2,7 @@ package pluginfx
 
 import (
 	"errors"
+	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/suite"
@@ -23,7 +24,37 @@ func (suite *ProvideSuite) testPGlobal() {
 		app = fxtest.New(
 			suite.T(),
 			P{
-				Path: "./sample.so",
+				Path: samplePath,
+				Symbols: Symbols{
+					Names: []interface{}{
+						"New",
+					},
+				},
+				Lifecycle: Lifecycle{
+					OnStart: "Initialize",
+					OnStop:  "Shutdown",
+				},
+			}.Provide(),
+			fx.Populate(&value, &p),
+		)
+	)
+
+	app.RequireStart()
+	app.RequireStop()
+
+	suite.Equal(expectedNewValue, value)
+	suite.NotNil(p)
+}
+
+func (suite *ProvideSuite) testPExpandEnv() {
+	var (
+		value float64
+		p     Plugin
+
+		app = fxtest.New(
+			suite.T(),
+			P{
+				Path: "${PWD}/" + samplePath,
 				Symbols: Symbols{
 					Names: []interface{}{
 						"New",
@@ -54,7 +85,7 @@ func (suite *ProvideSuite) testPAnonymous() {
 			P{
 				Anonymous: true,
 				Name:      "ShouldBeIgnored",
-				Path:      "./sample.so",
+				Path:      samplePath,
 				Symbols: Symbols{
 					Names: []interface{}{
 						"New",
@@ -91,7 +122,7 @@ func (suite *ProvideSuite) testPNamed() {
 			suite.T(),
 			P{
 				Name: "MyPlugin",
-				Path: "./sample.so",
+				Path: samplePath,
 				Symbols: Symbols{
 					Names: []interface{}{
 						"New",
@@ -127,7 +158,7 @@ func (suite *ProvideSuite) testPGroup() {
 			suite.T(),
 			P{
 				Group: "plugins",
-				Path:  "./sample.so",
+				Path:  samplePath,
 				Symbols: Symbols{
 					Names: []interface{}{
 						"New",
@@ -176,10 +207,132 @@ func (suite *ProvideSuite) testPAnonymousError() {
 
 func (suite *ProvideSuite) TestP() {
 	suite.Run("Global", suite.testPGlobal)
+	suite.Run("ExpandEnv", suite.testPExpandEnv)
 	suite.Run("Anonymous", suite.testPAnonymous)
 	suite.Run("Named", suite.testPNamed)
 	suite.Run("Group", suite.testPGroup)
 	suite.Run("AnonymousError", suite.testPAnonymousError)
+}
+
+func (suite *ProvideSuite) testSAnonymous() {
+	var (
+		value float64
+
+		app = fxtest.New(
+			suite.T(),
+			S{
+				Paths: []string{samplePath},
+				Symbols: Symbols{
+					Names: []interface{}{
+						"New",
+					},
+				},
+				Lifecycle: Lifecycle{
+					OnStart: "Initialize",
+					OnStop:  "Shutdown",
+				},
+			}.Provide(),
+			fx.Populate(&value),
+		)
+	)
+
+	app.RequireStart()
+	app.RequireStop()
+
+	suite.Equal(expectedNewValue, value)
+}
+
+func (suite *ProvideSuite) testSGroup() {
+	var (
+		value float64
+
+		app = fxtest.New(
+			suite.T(),
+			S{
+				Group: "plugins",
+				Paths: []string{samplePath},
+				Symbols: Symbols{
+					Names: []interface{}{
+						"New",
+					},
+				},
+				Lifecycle: Lifecycle{
+					OnStart: "Initialize",
+					OnStop:  "Shutdown",
+				},
+			}.Provide(),
+			fx.Populate(&value),
+			fx.Invoke(
+				func(in struct {
+					fx.In
+					Plugins []Plugin `group:"plugins"`
+				}) {
+					suite.Len(in.Plugins, 1)
+				},
+			),
+		)
+	)
+
+	app.RequireStart()
+	app.RequireStop()
+
+	suite.Equal(expectedNewValue, value)
+}
+
+func (suite *ProvideSuite) testSExpandEnv() {
+	var (
+		value float64
+
+		app = fxtest.New(
+			suite.T(),
+			S{
+				Group: "plugins",
+				Paths: []string{"${PWD}/*.so"},
+				Symbols: Symbols{
+					Names: []interface{}{
+						"New",
+					},
+				},
+				Lifecycle: Lifecycle{
+					OnStart: "Initialize",
+					OnStop:  "Shutdown",
+				},
+			}.Provide(),
+			fx.Populate(&value),
+			fx.Invoke(
+				func(in struct {
+					fx.In
+					Plugins []Plugin `group:"plugins"`
+				}) {
+					suite.Len(in.Plugins, 1)
+				},
+			),
+		)
+	)
+
+	app.RequireStart()
+	app.RequireStop()
+
+	suite.Equal(expectedNewValue, value)
+}
+
+func (suite *ProvideSuite) testSBadGlob() {
+	app := fx.New(
+		S{
+			Paths: []string{"["},
+		}.Provide(),
+	)
+
+	err := app.Err()
+	suite.Require().Error(err)
+	suite.True(errors.Is(err, filepath.ErrBadPattern))
+}
+
+func (suite *ProvideSuite) TestS() {
+	suite.Run("Anonymous", suite.testSAnonymous)
+	suite.Run("Group", suite.testSGroup)
+	suite.Run("ExpandEnv", suite.testSExpandEnv)
+	suite.Run("BadGlob", suite.testSBadGlob)
 }
 
 func TestProvide(t *testing.T) {